refactor: パッケージ固有型の整理
- packages/frontend/src/types/components.ts を作成
  - FileTreeProps, FileTreeItemProps, ProjectState を定義
- packages/frontend/src/types/file-tree.ts を更新
  - 非推奨マークを追加し、components.ts への再エクスポートに変更
- packages/frontend/src/stores/projectStore.ts を更新
  - 新しい型定義ファイルからインポートするように変更
- packages/backend/src/types/internal.ts を作成
  - バックエンド固有の内部型定義を集約

要件 3.1, 3.2 に対応
