+++ packages/frontend/src/stores/projectStore.ts
@@ -1,30 +1,7 @@
 import { create } from 'zustand';
 import { getProjects, addProject, removeProject, selectProject } from '@/services';
-import type { ProjectInfo } from '@kiro-lens/shared';
-import type { FileItem } from '@shared/types/file-tree';
-
-interface ProjectState {
-  // 状態
-  projects: ProjectInfo[];
-  currentProject?: ProjectInfo;
-  selectedFile?: FileItem;
-  isAddingProject: boolean;
-  isLoading: boolean;
-  error?: string;
-
-  // 計算プロパティ
-  hasKiroDir: boolean;
-
-  // アクション
-  loadProjects: () => Promise<void>;
-  addProject: (path: string) => Promise<void>;
-  removeProject: (projectId: string) => Promise<void>;
-  selectProject: (project: ProjectInfo) => Promise<void>;
-  setSelectedFile: (file?: FileItem) => void;
-  setAddingProjectMode: (isAdding: boolean) => void;
-  clearError: () => void;
-  setError: (error: string) => void;
-}
+import type { ProjectInfo, FileItem } from '@kiro-lens/shared';
+import type { ProjectState } from '@/types/components.js';
 
 // ApiClientインスタンスは不要 - 関数ベースAPIを使用
